# Innova.Application/DTOs/Comment/CommentDto.cs
namespace Innova.Application.DTOs.Comment;

public class CommentDto
{
    public int Id { get; set; }
    public string Content { get; set; } = string.Empty;
    public string AppUserId { get; set; } = string.Empty;
    public int IdeaId { get; set; }
    public int? ParentId { get; set; }
    public DateTime CreatedAt { get; set; }
    public DateTime? UpdatedAt { get; set; }
}